Course_Management_Django/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
mysqlclient
pymongo
djongo
django-environ
pytz